refactor: :recycle: refactor cleanup
