+++ packages/workspace-fusion/src/lib/utils/createFusionWorkspace.tsx
@@ -1,13 +1,10 @@
+import { PreConfig, UserConfig } from 'lib/types';
 import { FusionWorkspaceBuilder } from '../classes';
 import { FUSION_FILTER_PROVIDER_NAME } from './filter/addFilterContext';
 import { sortFusionTabs } from './fusionTabOrder';
 
-type UserConfig<TData extends Record<PropertyKey, unknown>> = (
-	builder: Omit<FusionWorkspaceBuilder<TData>, 'viewController'>
-) => FusionWorkspaceBuilder<TData>;
-
 export function createFusionWorkspace<TData extends Record<PropertyKey, unknown>>(
-	config: AppConfig<TData>,
+	config: PreConfig<TData>,
 	builderFunc: UserConfig<TData>
 ) {
 	const builder = builderFunc(new FusionWorkspaceBuilder(config.getIdentifier, config.appKey));
@@ -26,10 +23,3 @@ export function createFusionWorkspace<TData extends Record<PropertyKey, unknown>
 
 	return sortFusionTabs(viewController);
 }
-
-type AppConfig<TData> = {
-	getIdentifier: GetIdentifier<TData>;
-	appKey: string;
-};
-
-export type GetIdentifier<TData> = (item: TData) => string;
